refactor(repositories-config): rename 'config' to 'repositories-config'
